requirements.txt
cherrypy==3.6.0
nose==1.3.6
requests==2.6.0
numpy==1.9.0
scipy==0.13.3
scikit-learn==0.14.1
pystemmer==1.3.0
psycopg2==2.6
peewee==2.6.4
